Mega-commit
Snapshot of a bunch of stuff I've been working on.

Specifically I've been working on implementing a plugin system for parsing header fields.  It's about 70% done in my estimation - the groundwork is in place, but it isn't fully implemented yet and there are still errors in the implementation I'm working through.  Once complete, building new header parsers will be significantly easier as we will just write a plugin.